+++ dejafu/CHANGELOG.markdown
@@ -32,9 +32,12 @@ unreleased
 - The `Way` type is now abstract and exposes smart constructor functions:
     - `systematically`, corresponding to the old `Systematically`.
     - `randomly`, corresponding to the old `Randomly`,
+    - `uniformly`, a new uniform random (as opposed to weighted random) scheduler.
     - `swarmy`, corresponding to the old `Randomly` and specifying how many executions to use the
       same weights for.
-- The `sctRandom` function can now re-use the same weights for multiple executions.
+- A new `sctUniformRandom` function to do uniform (non-weighted) scheduling.
+- The `sctRandom` function is now called `sctWeightedRandom` and can now re-use the same weights for
+  multiple executions.
 
 ### Fixed