+++ hunit-dejafu/CHANGELOG.markdown
@@ -17,6 +17,7 @@ unreleased
 - Due to changes in dejafu, the `Way` type is now abstract and exposes smart constructor functions:
     - `systematically`, corresponding to the old `Systematically`.
     - `randomly`, corresponding to the old `Randomly`.
+    - `uniformly`, a new uniform random (as opposed to weighted random) scheduler.
     - `swarmy`, corresponding to the old `Randomly` and specifying how many executions to use the
       same weights for.
 - The `defaultWay`, `defaultMemType`, and `defaultBounds` values are all now re-exported.